refactor: change dtos package path
