--- src/main/java/br/com/zup/taxes/controllers/dto/TaxDto.java
@@ -1,18 +1,0 @@
-package br.com.zup.taxes.controllers.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class TaxDto {
-    private String name;
-    private String description;
-    private Double aliquot;
-}
