+++ src/main/java/br/com/zup/taxes/dtos/TaxDto.java
@@ -1,4 +1,4 @@
-package br.com.zup.taxes.controllers.dto;
+package br.com.zup.taxes.dtos;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -11,8 +11,7 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class TaxResponseDto {
-    private Long id;
+public class TaxDto {
     private String name;
     private String description;
     private Double aliquot;
