+++ src/main/java/br/com/zup/taxes/repositories/RoleRepository.java
@@ -1,6 +1,6 @@
 package br.com.zup.taxes.repositories;
 
-import br.com.zup.taxes.controllers.dto.RoleEnum;
+import br.com.zup.taxes.dtos.RoleEnum;
 
 public interface RoleRepository {
     boolean existsByName(String name);
