+++ src/main/java/br/com/zup/taxes/repositories/TaxRepository.java
@@ -1,7 +1,7 @@
 package br.com.zup.taxes.repositories;
 
-import br.com.zup.taxes.controllers.dto.TaxDto;
-import br.com.zup.taxes.controllers.dto.TaxResponseDto;
+import br.com.zup.taxes.dtos.TaxDto;
+import br.com.zup.taxes.dtos.TaxResponseDto;
 
 import java.util.List;
