+++ src/main/java/br/com/zup/taxes/services/TaxServiceImpl.java
@@ -1,7 +1,7 @@
 package br.com.zup.taxes.services;
 
-import br.com.zup.taxes.controllers.dto.TaxDto;
-import br.com.zup.taxes.controllers.dto.TaxResponseDto;
+import br.com.zup.taxes.dtos.TaxDto;
+import br.com.zup.taxes.dtos.TaxResponseDto;
 import br.com.zup.taxes.repositories.TaxRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
